Датчик тмпературы теперь на i2c - работает. Датчик освещенности пока такой же. Добавил отдельне модули на temp и illum. Добавлен канал ацп для напряжения и тока. Добавлен канал цап для тока. Написана функция мат модели. Включен fpu.

+++ Core/Inc/logger.h
@@ -7,7 +7,7 @@ typedef struct
 {
     uint8_t prxA;
     uint8_t prxB;
-    uint16_t temp;
+    float temp;
     uint16_t illum;
     uint8_t crc;
